+++ NEWS.md
@@ -1,3 +1,10 @@
+# coveffectsplot 0.0.2
+
+* Removed reference to the old name of the package
+* updated vignette and docs
+* tweaked the order of the shape legends and height of vertical dodging
+
+
 # coveffectsplot 0.0.1
 
 * Initial Release of coveffectsplot